Allow customize which type of body can be logged based on Content-Type
Allow customize which type of body can be logged based on Content-Type. 

The client can define its own list of Content-Type want to be list, default only log text and json.

The old implementation may log binary content, for example I have multipart content type, the first part is text, the second part is a binary file, it will log the binary in this case. With this new enhancement it will not log multipart content.